+++ CHANGELOG.md
@@ -11,6 +11,12 @@ Added / Changed / Deprecated / Fixed / Removed / Security
 
 > Corresponds to changes in the `develop` branch since the last release
 
+### Added
+
+#### org.ojalgo.matrix
+
+- Added the ability to specify the initial capacity of a `SparseStore`.
+
 ### Changed
 
 #### org.ojalgo.structure